SDK-887: Back-port changes for Drupal CI

+++ yoti/tests/src/Functional/YotiBrowserTestBase.php
@@ -10,7 +10,7 @@ use Drupal\yoti\YotiHelper;
  *
  * @group yoti
  */
-class YotiBrowserTestBase extends BrowserTestBase {
+abstract class YotiBrowserTestBase extends BrowserTestBase {
 
   /**
    * Linked User.
@@ -80,7 +80,9 @@ class YotiBrowserTestBase extends BrowserTestBase {
       mkdir(YotiHelper::uploadDir(), 0777, TRUE);
     }
     $this->selfieFilePath = YotiHelper::uploadDir() . DIRECTORY_SEPARATOR . 'test_selfie.jpg';
-    file_put_contents($this->selfieFilePath, 'test_selfie_contents');
+    if (!is_file($this->selfieFilePath)) {
+      file_put_contents($this->selfieFilePath, 'test_selfie_contents');
+    }
     $user_data[YotiHelper::ATTR_SELFIE_FILE_NAME] = basename($this->selfieFilePath);
 
     \Drupal::database()->insert(YotiHelper::YOTI_USER_TABLE_NAME)->fields([
@@ -90,21 +92,6 @@ class YotiBrowserTestBase extends BrowserTestBase {
     ])->execute();
 
     return $linkedUser;
-  }
-
-  /**
-   * Teardown Tests.
-   */
-  public function teardown() {
-    // Cleanup selfie.
-    if (is_file($this->selfieFilePath)) {
-      unlink($this->selfieFilePath);
-    }
-
-    // Cleanup private directory.
-    rmdir(YotiHelper::uploadDir());
-
-    parent::teardown();
   }
 
 }
